# perf: remove extra highlight calls